+++ app.py
@@ -3,16 +3,12 @@ import googleapiclient.discovery
 import streamlit.components.v1 as components
 import utils
 import pandas as pd
-import urllib.parse, urllib.request
-from urllib import parse
-from youtube_transcript_api import YouTubeTranscriptApi
-import json
-import queue
 import re
 import os
 import pickle
 from pyvis.network import Network
 from datetime import datetime, timedelta
+from Script_Exctractor import Script_Exctractor
 
 # Default values for options
 NUM_OF_VIDEOS = 10
@@ -44,101 +40,6 @@ with open("./data/new_learning_list.pkl", "rb") as file:
 with open("./data/selected_video.pkl", "rb") as file:
     selected_video = pickle.load(file)
 
-class Script_Exctractor:
-    def __init__(self,vid,setTime):
-        self.vid = vid
-        self.scriptData:list = []
-        self.setTime = setTime
-        self.wikiUserKey = "ecnkjsfhuuvmjfxdvdziepjwznhwdc"
-
-    # youtube script extraction
-    def Extract(self):
-        # Preprocessing URL for obtaining YouTube script
-        parsedUrl = parse.urlparse(self.vid)
-        vid = parse.parse_qs(parsedUrl.query)['v'][0]
-        languages = ['en','en-US']
-        str:list = YouTubeTranscriptApi.get_transcript(vid,languages)
-
-        # Splitting stored JSON information into time intervals (timeSet).
-        ret = queue.Queue()
-        nowSec = self.setTime
-        sentence = ''
-        for st in str:
-            if st['start'] >= nowSec:
-                ret.put(sentence)
-                sentence = ''
-                nowSec += self.setTime
-                if frontSt['start'] + frontSt['duration'] >= nowSec:
-                    sentence += frontSt['text'] + '\n'
-            sentence += st['text'] + '\n'
-            frontSt = st
-        ret.put(sentence)
-        while not ret.empty():
-            self.scriptData.append(ret.get())
-
-        for i in range(len(self.scriptData)):
-            text = self.scriptData[i].replace(u'\xa0', u' ').replace(u'\n',u' ').replace(u'  ',u' ')
-            self.scriptData[i] = text
-
-    def CallWikifier(self, text, lang="en", threshold=0.8, numberOfKCs=10):
-        # Prepare the URL.
-        data = urllib.parse.urlencode([
-                ("text", text), ("lang", lang),
-                ("userKey", self.wikiUserKey),
-                ("pageRankSqThreshold", "%g" % threshold),
-                ("applyPageRankSqThreshold", "true"),
-                ("nTopDfValuesToIgnore", "200"),
-                ("nWordsToIgnoreFromList", "200"),
-                ("wikiDataClasses", "false"),
-                ("wikiDataClassIds", "false"),
-                ("support", "false"),
-                ("ranges", "false"),
-                ("minLinkFrequency", "3"),
-                ("includeCosines", "false"),
-                ("maxMentionEntropy", "2")
-                ])
-        url = "http://www.wikifier.org/annotate-article"
-        # Call the Wikifier and read the response.
-        req = urllib.request.Request(url, data=data.encode("utf8"), method="POST")
-        with urllib.request.urlopen(req, timeout = 60) as f:
-            response = f.read()
-            response = json.loads(response.decode("utf8"))
-
-        sorted_data = sorted(response['annotations'], key=lambda x: x['pageRank'], reverse=True)
-        # Output the annotations.
-        num = 0
-        result = []
-        for annotation in sorted_data:
-            if num < numberOfKCs:
-                result.append({"title":annotation["title"],"url":annotation["url"],"pageRank":annotation["pageRank"]})
-
-            num += 1
-
-        res = result
-        result = []
-        return res
-
-    def UrltoWiki(self):
-        self.Extract()
-
-        number = 1
-        results = []
-        for text in self.scriptData:
-            print(f"{number} segemnt")
-            results.append(self.CallWikifier(text=text, numberOfKCs=NUM_OF_WORDS))
-            number += 1
-
-        wiki_data = pd.DataFrame()
-        seg_no = 1
-
-        for seg_item in results:
-            seg_df = pd.DataFrame(seg_item)
-            seg_df['seg_no'] = seg_no
-            seg_df['understand']=0
-            wiki_data = pd.concat([wiki_data,seg_df])
-            seg_no = seg_no + 1
-        wiki_data.index = range(len(wiki_data))
-        return wiki_data
 
 # Video recommendation using Jaccard similarity.
 class VideoRecommender:
@@ -204,6 +105,9 @@ class VideoRecommender:
         sorted_recommended_videos = [video for video, _ in recommended_videos]
         return sorted_recommended_videos
 
+###
+#   TODO: 추후 API KEY를 사용자에게 받아서 추가해야됨 
+###
 YOUTUBE_API_KEY = "AIzaSyCt74iOovLdzJMGCfsCAW4nAssQB8LJWo0"
 
 # API information
@@ -420,7 +324,7 @@ if search_button and user_input:
     # Wikification and ranking
     for index, video in enumerate(new_learning_list):
         try:
-            Scripts = Script_Exctractor(video.url, TIME_DIVISION)
+            Scripts = Script_Exctractor(video.url, TIME_DIVISION,NUM_OF_WORDS)
             video.segment = Scripts.UrltoWiki()
             # print(video.segment) # Display segments created for each video
         except Exception as ex:
